deps: move fetching into optional feature so that reqwest isn't required anymore
